Updated lmfit to latest upstream version 5.0
