+++ libraries/lmfit/Makefile.am
@@ -2,11 +2,11 @@ noinst_LTLIBRARIES += liblmfit.la
 
 liblmfit_la_CPPFLAGS = \
 	-I$(srcdir) \
-	-I$(top_srcdir)/libraries/lmfit/lmfit-3.2/lib \
+	-I$(top_srcdir)/libraries/lmfit/lmfit-5.0/lib \
         $(EXTRA_CFLAGS)
 
 liblmfit_la_SOURCES = \
-	libraries/lmfit/lmfit-3.2/lib/lmmin.c \
-	libraries/lmfit/lmfit-3.2/lib/lmmin.h \
-	libraries/lmfit/lmfit-3.2/lib/lmcurve.c \
-	libraries/lmfit/lmfit-3.2/lib/lmcurve.h
+	libraries/lmfit/lmfit-5.0/lib/lmmin.c \
+	libraries/lmfit/lmfit-5.0/lib/lmmin.h \
+	libraries/lmfit/lmfit-5.0/lib/lmcurve.c \
+	libraries/lmfit/lmfit-5.0/lib/lmcurve.h
