+++ libraries/lmfit/lmfit-5.0/COPYING
